Do not invoke the temperature setter when the temperature hasn't changed
This is an optimization. Doing a temperature set can be costly
(starting a process, in the case of the RedshiftTemperatureSetter).

We should avoid doing it every minute (or whatever the interval is).

This optimization has a shortcoming though.
If the user manually changes the temperature (by other means),
we would just say "same as last time" and not do anything.
That could also be a feature - preserving external changes,
until a new rule really takes effect and has meaningful changes to make.