Automatically activate on power resume
Putting a computer to sleep and resuming it later should
immediately trigger a temperature set.

When power resumes, a lot of time has potentially passed
and it's very likely that a new temperature needs to be set.
Instead of waiting for the next scheduled temperature set,
we'd better catch the resume event and activate immediately.